[#1211] Refactor tests.
